backend/requirements.txt
fastapi
pyjwt
passlib[bcrypt]
pydantic
starlette
httpx
uvicorn 
psycopg2-binary
sqlalchemy[asyncio] 
alembic 
python-dotenv
asyncpg
fastapi[standard]